% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{adobo_impute}
\alias{adobo_impute}
\title{Call cell types in a Seurat object using adobo, a Python package that calls
cell types using a marker gene list.}
\usage{
adobo_impute(
  st_obj,
  gene_symbols,
  cell_types,
  cluster_slot = NULL,
  out_cluster_slot = "adobo_clusters",
  symbol_switch = NULL,
  new_idents = FALSE
)
}
\arguments{
\item{st_obj}{Seurat object to process.}

\item{gene_symbols}{A vector of gene symbols corresponding to the marker
genes for each cluster.}

\item{cell_types}{A vector of cell types corresponding to the marker genes
for each cluster.}

\item{cluster_slot}{The slot name in the Seurat dataset corresponding to
cell identities or cluster names. The default behavior
is Idents(st_obj).}

\item{symbol_switch}{Whether to switch ENSEMBL to MGI symbols for "mouse"
or "human". Default is to not switch (None). Input
gene format must be gene symbols in that case.}

\item{new_idents}{Whether to set the adobo output results as new cell
idents. Default is FALSE.}
}
\value{
The modified Seurat object with adobo clusters.
}
\description{
Call cell types in a Seurat object using adobo, a Python package that calls
cell types using a marker gene list.
}
