% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{apply_ensembl_seurat}
\alias{apply_ensembl_seurat}
\title{Apply ENSEMBL naming to Seurat objects}
\usage{
apply_ensembl_seurat(st, mart = NULL, mapper_table = NULL)
}
\arguments{
\item{st}{The Seurat object. Make sure that the default assay is the one that
is to be processed.}

\item{mart}{If mapper_table is NULL, the mart to be used for fetching ENSEMBL IDs
and MGI symbols. Default is NULL. If mapper_table is not NULL, this
parameter will be ignored.}

\item{mapper_table}{The mapping table to use. Default is NULL; if a table is not
provided, then one will be created using the
\code{create_default_mapper_table} function.}
}
\value{
A Seurat object with renamed symbols in a new assay with 'ENSEMBL'
prepended to the old assay name. This new assay will be the
default.
}
\description{
Given a Seurat object with a specific name and assay, this function renames
the genes in the assay (assuming the gene names are MGI symbols) by
replacing them with their ENSEMBL IDs (ENSMUSG...).
}
