% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_if_data_exists}
\alias{check_if_data_exists}
\title{Check if dmSC datasets are already installed. Not exported.}
\usage{
check_if_data_exists(data_nm)
}
\value{
Whether a data name exists in the available search paths.
}
\description{
Check if dmSC datasets are already installed. Not exported.
}
