% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_data_folder}
\alias{create_data_folder}
\title{Create the data folder for holding dmSC datasets.}
\usage{
create_data_folder(verbose = FALSE)
}
\description{
Create the data folder for holding dmSC datasets.
}
