% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_default_mapper_table}
\alias{create_default_mapper_table}
\title{Default mapping table creation}
\usage{
create_default_mapper_table(mart = NULL, to_ensembl = FALSE)
}
\arguments{
\item{mart}{The mart to be used for fetching ENSEMBL IDs and MGI symbols.
Default is NULL. @param to_ensembl Whether ENSEMBL IDs should be
the keys (FALSE) or values (TRUE). Default is FALSE. @returns A list
containing the following: * A one-to-one mapping table based
on the description above ('mapper_table'), * A data.table containing the
duplicated keys that were removed ('duplicate_keys'), * A data.table
containing the duplicated values that were removed
('duplicate_values').}
}
\description{
Creates a default mapping table mapping between MGI symbols and ENSEMBL IDs.
Keys are in the first column, values are in the second. One-to-many and
one-to-one relations are tossed out to avoid problems with
preprocessing_function.
}
\details{
Currently creates a one-to-one mapping table based on Mus Musculus genes and
ENSEMBL version 102.

Note that we could allow one-to-one and many-to-one mappings, but NOT one-to-
many. However, the assumption toward the end of preprocessing_function would
be broken and we would need to write code that pools values with the same
ENSEMBL ID.
}
