% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{create_split_heatmap}
\alias{create_split_heatmap}
\title{Generate a grouped heatmap for a specific list of genes and corresponding
expression data. Just a wrapper around pheatmap::pheatmap.}
\usage{
create_split_heatmap(
  assay,
  assay_data_columns = NULL,
  type_column = "TYPE",
  symbol_column = "SYMBOL",
  gap_length = 3,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  border_color = NA,
  ...
)
}
\arguments{
\item{assay}{The assay to include. Should be a 2D table with a column of gene symbols and
at least one column of corresponding gene expression values for that symbol.
Can also include a 'type column', or a column including the categories under
which each gene falls (if you're trying to split a heatmap based on specific
categories of genes). Ideally a data.table or a data.frame.}

\item{assay_data_columns}{A vector of names specifying the data columns within assay.
The default (NULL) just takes all columns besides the symbol
and type columns.}

\item{type_column}{The name of the column of assay specifying the category that each gene falls into.
If NULL, then no types are assumed and the heatmap won't be split.
Default is "TYPE".}

\item{symbol_column}{The column of assay specifying gene symbols. Default is "SYMBOL".}

\item{gap_length}{The space between splits on the split heatmap. Default is 3.}

\item{cluster_rows}{Whether to cluster heatmap rows within splits. Default is FALSE.}

\item{cluster_cols}{Whether to cluster columns across all heatmap entries, disregarding
splits. Default is FALSE.}

\item{border_color}{The color of border to use on heatmaps. Default is NA, or no color.}

\item{...}{Other arguments to be passed into pheatmap::pheatmap.}
}
\value{
Heatmap object as output by pheatmap::pheatmap.
}
\description{
Generate a grouped heatmap for a specific list of genes and corresponding
expression data. Just a wrapper around pheatmap::pheatmap.
}
