% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{eliminate_mapper_all_duplicates}
\alias{eliminate_mapper_all_duplicates}
\title{Eliminate all duplicates in mapping table}
\usage{
eliminate_mapper_all_duplicates(mapper_table)
}
\arguments{
\item{mapper_table}{The mapper_table whose values are to be de-duplicated.}
}
\value{
A list containing the following:
* A deduplicated mapping table based on the description above
('mapper_table'),
* A data.table containing the duplicated keys that were removed
('duplicate_keys').
* A data.table containing the duplicated values that were removed
('duplicate_values').
}
\description{
Eliminates duplicate values in all columns (both keys and values) of a
mapper_table, returning both the de-duplicated mapper_table and data.frames
consisting of the removed duplicate keys/values (and the number of occurrences
for each duplicate key/value).
}
