% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{eliminate_mapper_key_duplicates}
\alias{eliminate_mapper_key_duplicates}
\title{Eliminate duplicates in mapping table keys}
\usage{
eliminate_mapper_key_duplicates(mapper_table)
}
\arguments{
\item{mapper_table}{The mapper_table whose keys are to be de-duplicated.}
}
\value{
A list containing the following:
* A deduplicated mapping table based on the description above
('mapper_table'),
* A data.table containing the duplicated keys that were removed
('duplicate_keys').
}
\description{
Eliminates duplicate values in the 'keys' (first column) of a mapper_table,
returning both the de-duplicated mapper_table and a data.frame consisting
of the removed duplicate keys (and the number of occurrences for each
duplicate key).
}
