% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_datapath}
\alias{get_datapath}
\title{View search locations for dmSC datasets.}
\usage{
get_datapath()
}
\value{
A vector describing the current search path.
}
\description{
View search locations for dmSC datasets.
}
