% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{impute_st}
\alias{impute_st}
\title{Impute cell types for Seurat object tgt_st with Seurat object ref_st.}
\usage{
impute_st(
  tgt_st,
  ref_st,
  out_nm = "singler_clusters",
  tgt_clusters_nm = NULL,
  ref_clusters_nm = NULL
)
}
\arguments{
\item{tgt_st}{Target Seurat object to impute cell types against.}

\item{ref_st}{Reference Seurat object against which to impute cell types.}

\item{out_nm}{Column to create imputed clusters under in tgt_st.}

\item{tgt_clusters_nm}{Column in tgt_st containing clusters. If NULL, Seurat::Idents(tgt_st) is used.}

\item{ref_clusters_nm}{Column in ref_st containing clusters. If NULL, Seurat::Idents(ref_st) is used.}
}
\value{
Target Seurat object with extra column containing imputed clusters.
}
\description{
Impute cell types for Seurat object tgt_st with Seurat object ref_st.
}
