% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{install_data}
\alias{install_data}
\title{Install dmSC datasets from local locations or online locations.}
\usage{
install_data(nm, ..., reinstall = FALSE, verbose = FALSE)
}
\arguments{
\item{nm}{Name of the dataset to install. Do not include extension.}

\item{...}{Names of additional datasets to install. Optional.}

\item{reinstall}{Whether to re-copy the dataset to your installed
data files if the requested dataset is installed.
Default FALSE.}

\item{verbose}{Whether to log messages. Default FALSE.}
}
\description{
Install dmSC datasets from local locations or online locations.
}
