% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ipa_dotplot}
\alias{ipa_dotplot}
\title{A simple function for creating dot plots for gene ontology (GO) or Ingenuity Pathway Analysis
(IPA) plots. Code taken from enrichplot: https://github.com/YuLab-SMU/enrichplot/blob/devel/R/dotplot.R#L200-L207}
\usage{
ipa_dotplot(
  tbl,
  z_score_colname,
  ratio_colname,
  neg_log_p_colname,
  x_axis_label_colname = NULL,
  y_axis_label_colname = NULL
)
}
\arguments{
\item{tbl}{A data.frame or similar containing gene symbols, z-scores, gene set ratios,
and -log10(p) values. Both of these should be output or obtainable from GO enrichment
analysis or IPA results.}

\item{z_score_colname}{Name of the column containing results for z-scores.}

\item{ratio_colname}{Name of the column containing gene set ratios.}

\item{neg_log_p_colname}{Name of the column containing -log10(p) values.}

\item{x_axis_label_colname}{Name of the column containing gene symbols for plotting along the
x-axis. This will create a horizontal plot. Note that either this
or y_axis_label_colname should be set but not both.}

\item{y_axis_label_colname}{Name of the column containing gene symbols for potting along the
y-axis. This will create a vertical plot. Note that either this
or y_axis_label_colname should be set but not both.}
}
\value{
A ggplot2-compatible dotplot summarizing the input data.
}
\description{
A simple function for creating dot plots for gene ontology (GO) or Ingenuity Pathway Analysis
(IPA) plots. Code taken from enrichplot: https://github.com/YuLab-SMU/enrichplot/blob/devel/R/dotplot.R#L200-L207
}
