% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessing_function}
\alias{preprocessing_function}
\title{Standard preprocessing function for single-cell data}
\usage{
preprocessing_function(st, mart = NULL, mapper_table = NULL)
}
\arguments{
\item{st}{The Seurat single-cell file to be preprocessed. Read above for how to
ensure this file is structured.}

\item{mart}{If mapper_table is NULL, the mart to be used for fetching ENSEMBL IDs
and MGI symbols. Default is NULL. If mapper_table is not NULL, this
parameter will be ignored.}

\item{mapper_table}{The mapping table to use. Default is NULL; if a table is not
provided, then one will be created using the
\code{create_default_mapper_table} function.}
}
\value{
The preprocessed Seurat single-cell file. A new assay will be created with
the key 'ENSEMBL' appended onto the end of the default assay of the input
Seurat file.
}
\description{
DEPRECATED: Recommended to just do preprocessing yourself so you have
more granular control over parameters. See the source code of this function
for an idea/starting point if still interested.
}
\details{
TO DO:
\itemize{
\item Make mitochondrial gene identification case-insensitive.
\item Manual assay specification in addition to just using the default assay.
\item Preserve variable features in the ENSEMBL assay since we just change
the gene names.
\item It seems that this function does not work when multiple assays are present.
Fix that.
}

Preprocess a Seurat single-cell dataset using the following techniques:
\itemize{
\item Remove genes that are expressed in fewer than 5\% of cells.
\item Remove genes that are not expressed or have no variance in expression across
all cells.
\item Remove cells with UMI counts, feature counts, mitochondrial gene percent, and
ribosomal gene percent outside 3 mean absolute deviations from the median.
\item Remove cells with no ENSEMBL ID counterpart.
}

Note that in the Seurat dataset, mitochondrial genes must begin with 'mt-' and
ribosomal genes must begin with 'Rp'. The genes also cannot be ENSEMBL IDs to
begin with. The gene name requirements are case-sensitive. Some mitochondrial
genes may begin with CAPITAL 'MT-' based on the source; this will cause problems.
Also ensure that the assay of interest in the Seurat object is marked as the
default assay, since this will be the assay that will be used.
}
