% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_datapath}
\alias{set_datapath}
\title{Set the search locations for dmSC datasets.}
\usage{
set_datapath(...)
}
\arguments{
\item{...}{An array of paths to search for dmSC datasets.}
}
\description{
Set the search locations for dmSC datasets.
}
