% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot a nice-looking volcano plot with minimal effort. Thanks, Elvis.}
\usage{
volcano_plot(
  df,
  logfc_column,
  logfc_cutoff,
  padj_column,
  padj_cutoff = 0.05,
  point_shape = 21,
  hline_linetype = "dashed",
  vlines_linetype = "dashed"
)
}
\arguments{
\item{df}{The data frame to input. This would ideally be the output of a program like
DESeq2, containing logarithmic fold changes and adjusted p-values.}

\item{logfc_column}{The column name containing the logarithmic fold-changes.}

\item{logfc_cutoff}{The logarithmic fold-change cutoff that determines "high" or "low" values.}

\item{padj_column}{The column name containing adjusted p-values.}

\item{padj_cutoff}{The default p-value cutoff. Default is 0.05.}

\item{point_shape}{The shape of the points to plot. Default is 21, corrsponding to circles.
See ggplot2 for more information.}

\item{hline_linetype}{The type of boundary to plot between significant and nonsignificant points.
Default is a dashed line.}

\item{vlines_linetype}{The type of boundary to plot between high/low points and points with lower
logarithmic fold changes. Default is a dashed line.}
}
\value{
A ggplot2-compatible volcano plot.
}
\description{
Plot a nice-looking volcano plot with minimal effort. Thanks, Elvis.
}
